requirements.txt
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
rich==14.1.0
rich-toolkit==0.15.0
starlette==0.47.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
